fixes #4621
