+++ Rubberduck.Core/AutoComplete/Service/SelfClosingPairHandler.cs
@@ -84,26 +84,38 @@ namespace Rubberduck.AutoComplete.Service
         {
             if (!original.CaretPosition.IsSingleCharacter)
             {
-                // todo: WrapSelection?
+                // todo: WrapSelectionWith(pair)?
                 result = null;
                 return false;
             }
 
-            var isPresent = original.CaretLine.EndsWith($"{pair.OpeningChar}{pair.ClosingChar}");
-
+            // if executing the SCP against the original code yields no result, we need to bail out.
             if (!_scpService.Execute(pair, original, e.Character, out result))
             {
                 return false;
             }
 
+            // let the VBE alter the original code if it wants to, then work with the prettified code.
             var prettified = CodePaneHandler.Prettify(e.Module, original);
+
+            var isPresent = original.CaretLine.EndsWith($"{pair.OpeningChar}{pair.ClosingChar}");
             if (!isPresent && original.CaretLine.Length + 2 == prettified.CaretLine.Length &&
                 prettified.CaretLine.EndsWith($"{pair.OpeningChar}{pair.ClosingChar}"))
             {
                 // prettifier just added the pair for us; likely a Sub or Function statement.
-                prettified = original; // pretend this didn't happen. note: probably breaks if original has extra whitespace.
+                prettified = original; // pretend this didn't happen; we need to work out the caret position anyway.
             }
 
+            if (prettified.CaretLine.Length == 0)
+            {
+                // prettifier destroyed the indent. need to reinstate it now.
+                prettified = prettified.ReplaceLine(
+                    index:prettified.CaretPosition.StartLine,
+                    content:new string(' ', original.CaretLine.TakeWhile(c => c == ' ').Count())
+                );
+            }
+
+            // if executing the SCP against the prettified code yields no result, we need to bail out.
             if (!_scpService.Execute(pair, prettified, e.Character, out result))
             {
                 return false;
@@ -117,7 +129,7 @@ namespace Rubberduck.AutoComplete.Service
                 result = null;
                 return false;
             }
-
+            
             var currentLine = reprettified.Lines[reprettified.CaretPosition.StartLine];
             if (!string.IsNullOrWhiteSpace(currentLine) &&
                 currentLine.EndsWith(" ") &&
@@ -130,7 +142,7 @@ namespace Rubberduck.AutoComplete.Service
                 e.Character == pair.OpeningChar &&
                 !result.CaretLine.EndsWith($"{pair.OpeningChar}{pair.ClosingChar}"))
             {
-                // VBE eats it. bail out but still swallow the keypress.
+                // VBE eats it. bail out but still swallow the keypress; we already prettified the opening character into the editor.
                 e.Handled = true;
                 result = null;
                 return false;
