+++ RubberduckTests/AutoComplete/SelfClosingPairCompletionServiceTests.cs
